[Fix] - D07 EX01 Changed array len to sizeT

+++ d07/ex01/iter.cpp
@@ -6,7 +6,7 @@
 /*   By: pbie <pbie@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/04 16:47:49 by pbie              #+#    #+#             */
-/*   Updated: 2019/04/04 18:02:55 by pbie             ###   ########.fr       */
+/*   Updated: 2019/04/04 18:06:51 by pbie             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,11 +27,11 @@ class Test {
 std::ostream & operator<<( std::ostream & o, Test const & rhs ) { o << rhs.get(); return o; }
 
 template <typename T>
-void iter(T *array, int len, void(*f)(T const &))
+void iter(T *array, size_t len, void(*f)(T const &))
 {
 	if (array)
 	{
-		for(int i = 0; i < len; i++)
+		for(size_t i = 0; i < len; i++)
 		{
 			f(array[i]);
 		}
